first compile in 15 years. fixes to get it to compile without warnings or errors

+++ src/mwgwd/callstore.c
@@ -604,7 +604,7 @@ int storeIPCCall(Call * cmsg, Connection *conn)
 {
    struct PendingCall * pc;
    char * data;
-   int datalen, rc;
+   size_t datalen, rc;
 
    ENTER();
    storeLockCall();
